chore: Merge the .issue command into the github cog
